feat: Add solution for modules/esm.mjs
