+++ src/modules/test.js
@@ -1,0 +1,8 @@
+import { unknownObject, myServer } from "./esm.mjs";
+console.log(
+  "unknownObject:",
+  unknownObject,
+  "\n-----\n",
+  "myServer:",
+  myServer
+);
